requirements.txt
torchvision==0.11.2+cu113
numpy==1.18.0
Pillow==8.3.1
imageio==2.9.0
scipy==1.6.3
torch==1.10.1+cu113
tqdm==4.48.2
tensorboard==2.4.0
